requirements.txt
fastapi~=0.65.2
pytest==6.2.4
pyyaml==5.4.1
py~=1.10.0
pip~=21.1.2
wheel~=0.36.2
attrs~=21.2.0
toml~=0.10.2
MarkupSafe~=2.0.1
Werkzeug~=2.0.1
Jinja2~=3.0.1
click~=8.0.1
itsdangerous~=2.0.1
pluggy~=0.13.1
iniconfig~=1.1.1
pyparsing~=2.4.7
setuptools~=56.0.0
uvicorn~=0.14.0